Add confirmation to deleting PA dashboard (#4851)
* Add confirmation to deleting PA dashboard

* Add default confirmation content

* PR feedback

+++ packages/front-end/pages/product-analytics/dashboards/index.tsx
@@ -461,11 +461,22 @@ export default function DashboardsPage() {
                                       <DropdownMenuItem
                                         disabled={!canDelete}
                                         color="red"
-                                        onClick={async () => {
-                                          await apiCall(`/dashboards/${d.id}`, {
-                                            method: "DELETE",
-                                          });
-                                          mutateDashboards();
+                                        confirmation={{
+                                          confirmationTitle: (
+                                            <span>
+                                              Delete Dashboard <i>{d.title}</i>?
+                                            </span>
+                                          ),
+                                          cta: "Delete",
+                                          submit: async () => {
+                                            await apiCall(
+                                              `/dashboards/${d.id}`,
+                                              {
+                                                method: "DELETE",
+                                              },
+                                            );
+                                            mutateDashboards();
+                                          },
                                         }}
                                       >
                                         Delete
